+++ src/quality.rs
@@ -118,7 +118,7 @@ mod tests {
 
     #[test]
     fn converts_to_primitive() {
-        assert_eq!(11 as u32, Quality::Strange as u32);
+        assert_eq!(11_u32, Quality::Strange as u32);
     }
     
     #[test]
